requirements.txt
azure-ai-documentintelligence
azure-identity
openai
python-dotenv
openpyxl
streamlit
pandas
Pillow
PyMuPDF
requests
